requirements.txt
asgiref==3.2.7
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-formtools==2.2
django-heroku==0.3.1
django-weasyprint==0.5.5
gunicorn==20.0.4
html5lib==1.0.1
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
Pyphen==0.9.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1
